requirements.txt
flask==2.0.1
tensorflow==2.6.0
keras==2.6.0
numpy==1.19.5
opencv-python==4.5.3.56
pillow==9.5.0
pyttsx3==2.90
simple-websocket==0.10.0
gevent==21.8.0
gevent-websocket==0.10.1
werkzeug==2.0.1
h5py~=3.1.0
